+++ backend/internal/cocktaildb/conversions.go
@@ -3,10 +3,10 @@ package cocktaildb
 const mlPerOz = 29.57353193
 const mlPerShot = 45
 
-func OzToML(oz float32) float32 {
+func OzToML(oz float64) float64 {
 	return oz * mlPerOz
 }
 
-func ShotToML(oz float32) float32 {
+func ShotToML(oz float64) float64 {
 	return oz * mlPerShot
 }
